Bug fix - require `hexbin`
Turns out there is a cryptic bug that only shows up on R 4.1.2 and Ubuntu 20.04, in which `stat_summary_hex()` fails with an unhelpful warning. This is a big deal when we moved over to Docker. Solution is to require `hexbin` is installed and loaded. See this thread for more details: https://bytemeta.vip/repo/GreenleafLab/ArchR/issues/1292